fix: changes size of "Size" field from 4 to 8
While inserting into the database I found out that the mode of directory
in WIndows is stored in an uint32, meaning an unsigned integer of 32
bits.

When I was inserting a size for a directory in my database, postgres
would try to insert a 4 billion field into an signed integer of 32 bits
which would lead to an error.

That is why I changed the size of the "Size" field from "files" table
to be a bigint, this way I can insert uint32 bits with no problem.

However, this will require when taking out the "size" field from the
database an conversion back to an uint32.

This can't be escaped since postgres doesn't provide support for
unsigned integers.

+++ database/migrate/migrations/000001_create_files_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS files (
     name text NOT NULL,
     size bigint,
     is_dir bool,
-    mode integer,
+    mode bigint,
     extension text,
     updated_at TIMESTAMP NOT NULL,
     content text,
